fix(security): encrypt sensitive data, fix TLS bypass, sanitize errors
- Route bangumiAccessToken and webDavPassword through FlutterSecureStorage
  with one-time plaintext Hive migration on startup
- Replace hardcoded allowBadCertificates=true with user-configurable setting
  (default false, add toggle in proxy settings page)
- Remove e.toString() from user-facing error messages in video controller
- Extract SettingsKeys to settings_keys.dart (required by new setting keys)

Add flutter_secure_storage ^10.3.1 dependency.
