+++ linux/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   desktop_webview_window
   dynamic_color
   file_selector_linux
+  flutter_secure_storage_linux
   flutter_volume_controller
   media_kit_libs_linux
   media_kit_video
