+++ pubspec.yaml
@@ -95,6 +95,7 @@ dependencies:
   antlr4: ^4.13.2
   fl_chart: ^1.1.1
   flutter_foreground_task: ^9.0.0
+  flutter_secure_storage: ^10.3.1
   skeletonizer: ^2.1.2
   flutter_inappwebview_platform_interface: ^1.3.0+1
   flutter_inappwebview_ios: ^1.1.2
@@ -179,6 +180,7 @@ dev_dependencies:
   flutter_launcher_icons: ^0.14.3
   flutter_native_splash: ^2.4.3
   msix: ^3.16.12
+  mocktail: ^1.0.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
